rename TIFJobUpdateService to ThreatIntelFeedIndexService
